Add contextualised section functionality
